+++ examples/xegpu/softmax.py
@@ -13,35 +13,16 @@ from functools import cached_property
 import numpy as np
 from mlir import ir
 from mlir.execution_engine import ExecutionEngine
-from mlir.dialects import linalg, gpu, bufferization, arith, tensor, func, math
-from mlir.dialects import transform
-from mlir.dialects.transform import structured, loop, xegpu
 
 from lighthouse import dialects as lh_dialects
 from lighthouse.workload import benchmark, get_bench_wrapper_schedule
 from lighthouse.utils.memref import to_ctype as memref_to_ctype
 from lighthouse.utils.numpy import numpy_to_ctype
-from lighthouse.utils.mlir import func_cif
 from lighthouse.ingress.mlir_gen import get_mlir_elem_type
-from lighthouse.ingress.mlir_gen.gpu_utils import emit_gpu_util_funcs, emit_buf_to_tensor
-from lighthouse.pipeline.helper import (
-    apply_registered_pass,
-    canonicalize,
-    match,
-)
-from mlir.dialects.transform import bufferization as transform_bufferization
-from mlir.dialects.bufferization import LayoutMapOption
+from lighthouse.ingress.mlir_gen.gpu_softmax_payload import generate_gpu_softmax_payload
+from lighthouse.schedule.xegpu.softmax_schedule import get_softmax_schedule_module
 
 from xegpu_workload import XeGPUWorkload
-
-def match_and_split(*args, nhandles=1, **kwargs):
-    """Helper function that splits matched handles."""
-    matched = match(*args, **kwargs)
-    anytype = transform.AnyOpType.get()
-    matched_ops = transform.split_handle((anytype,) * nhandles, matched)
-    if nhandles == 1:
-        matched_ops = [matched_ops]
-    return matched_ops
 
 
 def softmax_complexity(M: int, N: int, nbytes: int):
@@ -180,262 +161,25 @@ class XeGPUSoftmax(XeGPUWorkload):
 
     def payload_module(self) -> ir.Module:
         """Generate MLIR module for softmax payload."""
-        mod = ir.Module.create()
         dtype = get_mlir_elem_type(self.dtype_str)
-        memref_t = ir.MemRefType.get(self.shape, dtype)
-        
-        with ir.InsertionPoint(mod.body):
-            # Function signature: payload(output, input)
-            @func_cif(memref_t, memref_t, name=self.payload_function_name)
-            def payload(output, input_arg):
-                # Convert memrefs to tensors
-                output_tensor = emit_buf_to_tensor(output, restrict=True, writable=True)
-                input_tensor = emit_buf_to_tensor(input_arg, restrict=True)
-                
-                M, N = self.shape
-                
-                # Define affine maps for indexing
-                # #map = affine_map<(d0, d1) -> (d0, d1)>  (identity 2D)
-                # #map1 = affine_map<(d0, d1) -> (d0)>     (broadcast/reduce along d1)
-                d0 = ir.AffineDimExpr.get(0)
-                d1 = ir.AffineDimExpr.get(1)
-                map_2d = ir.AffineMap.get(2, 0, [d0, d1])
-                map_1d = ir.AffineMap.get(2, 0, [d0])
-                
-                # Step 1: Find max - linalg.generic reduction
-                neg_inf = arith.constant(dtype, float('-inf'))
-                max_init = tensor.empty((M,), dtype)
-                max_filled = linalg.fill(neg_inf, outs=[max_init])
-                
-                @linalg.generic(
-                    [input_tensor],  # inputs
-                    [max_filled],  # outputs
-                    [map_2d, map_1d],  # indexing_maps
-                    [linalg.IteratorType.parallel, linalg.IteratorType.reduction],  # iterator_types
-                )
-                def row_max(in_val, acc):
-                    return arith.maximumf(in_val, acc)
-                
-                # Step 2: Subtract max (broadcast) - linalg.generic elementwise
-                output_init = tensor.empty((M, N), dtype)
-                
-                @linalg.generic(
-                    [input_tensor, row_max],  # inputs
-                    [output_init],  # outputs
-                    [map_2d, map_1d, map_2d],  # indexing_maps
-                    [linalg.IteratorType.parallel, linalg.IteratorType.parallel],  # iterator_types
-                )
-                def shifted(in_val, max_val, out):
-                    return arith.subf(in_val, max_val)
-                
-                # Step 3: Compute exp - linalg.generic elementwise
-                @linalg.generic(
-                    [shifted],  # inputs
-                    [output_init],  # outputs
-                    [map_2d, map_2d],  # indexing_maps
-                    [linalg.IteratorType.parallel, linalg.IteratorType.parallel],  # iterator_types
-                )
-                def exp_vals(in_val, out):
-                    return math.exp(in_val)
-                
-                # Step 4: Sum exp values - linalg.generic reduction
-                # Create collapsed tensor for sum init
-                # sum_init_2d = tensor.empty((M, 1), dtype)
-                sum_init = tensor.empty((M,), dtype)
-                # tensor.CollapseShapeOp(sum_init, sum_init_2d, [[0, 1]])
-
-                
-                zero = arith.constant(dtype, 0.0)
-                sum_filled = linalg.fill(zero, outs=[sum_init])
-                
-                @linalg.generic(
-                    [exp_vals],  # inputs
-                    [sum_filled],  # outputs
-                    [map_2d, map_1d],  # indexing_maps
-                    [linalg.IteratorType.parallel, linalg.IteratorType.reduction],  # iterator_types
-                )
-                def row_sum(in_val, acc):
-                    return arith.addf(in_val, acc)
-                
-                # Step 5: Divide by sum (broadcast) - linalg.generic elementwise
-                @linalg.generic(
-                    [exp_vals, row_sum],  # inputs
-                    [output_init],  # outputs
-                    [map_2d, map_1d, map_2d],  # indexing_maps
-                    [linalg.IteratorType.parallel, linalg.IteratorType.parallel],  # iterator_types
-                )
-                def result(exp_val, sum_val, out):
-                    return arith.divf(exp_val, sum_val)
-                
-                # Materialize result back to output memref
-                bufferization.materialize_in_destination(
-                    None, result, output, restrict=True, writable=True
-                )
-
-            # Emit utility functions for GPU memory management
-            emit_gpu_util_funcs(dtype, rank=2)
-
-        return mod
+        return generate_gpu_softmax_payload(
+            func_name=self.payload_function_name,
+            M=self.M,
+            N=self.N,
+            dtype=dtype,
+        )
 
     def schedule_modules(
         self, stop_at_stage: Optional[str] = None, parameters: Optional[dict] = None
     ) -> list[ir.Module]:
-        """
-        Generate transform schedule for softmax.
-        
-        For now, returns an empty schedule. In the future, this would contain
-        tiling, vectorization, and XeGPU-specific lowering transformations.
-        """
-        # TODO: Implement proper transform schedule
-        # For now, create a minimal schedule that prints the last linalg operation
-        mod = ir.Module.create()
-        mod.operation.attributes["transform.with_named_sequence"] = ir.UnitAttr.get()
-        
-        with ir.InsertionPoint(mod.body):
-
-            
-            # Create a transform sequence with proper signature
-            named_sequence = transform.named_sequence(
-                "__transform_main",
-                [transform.AnyOpType.get()],  # input: module
-                [],  # no outputs
-                arg_attrs=[{"transform.readonly": ir.UnitAttr.get()}]
-            )
-            
-            with ir.InsertionPoint(named_sequence.body):
-                # match the payload module
-                anytype = transform.AnyOpType.get()
-                func = match(named_sequence.bodyTarget, ops={"func.func"})
-                payload_mod = transform.get_parent_op(
-                    anytype,
-                    func,
-                    op_name="builtin.module",
-                    deduplicate=True,
-                )
-                
-                # Match all linalg.generic operations
-                # We have 5 generic ops in softmax: max, sub, exp, sum, div
-                generic_ops = structured.structured_match(
-                    transform.AnyOpType.get(),
-                    payload_mod,
-                    ops=["linalg.generic", "linalg.fill"]
-                )
-                
-                # Split the handle into individual operation handles
-                # For softmax, we have 7 operations
-                anytype = transform.AnyOpType.get()
-                split_ops = transform.split_handle(
-                    (anytype, anytype, anytype, anytype, anytype, anytype, anytype),  # 7 result types
-                    generic_ops
-                )
-                
-                # Reverse split_ops to have operations in reverse order
-                split_ops = list(reversed(split_ops))
-                
-                # The first operation (after reversal) is the division - this is the consumer
-                last_op = split_ops[0]
-
-                # Print the last operation before tiling
-                # transform.print_(target=last_op, name="last_linalg_generic_before_tiling")
-
-                # Tile the last operation using tile_using_forall
-                # Tile sizes: [64, 64] for the two parallel dimensions (M, N)
-                tiled_op, for_op = structured.structured_tile_using_forall(
-                    anytype, anytype,
-                    last_op,
-                    num_threads=[],
-                    tile_sizes=[],
-                    static_tile_sizes=(parameters["wg_rows"],),
-                )
-
-                # Fuse the producer operations into the forall loop
-                # Iterate through remaining operations (already in reverse order)
-                current_forall = for_op
-                for producer_op in split_ops[1:]:
-                    fused_op, current_forall = structured.structured_fuse_into_containing_op(
-                        anytype, anytype,
-                        producer_op,
-                        current_forall
-                    )
-                    
-                func = transform.get_parent_op(
-                    anytype,
-                    current_forall,
-                    op_name="func.func",
-                    deduplicate=True,
-                )
-                transform.apply_cse(func)
-                canonicalize(func)
-                
-                func = structured.VectorizeChildrenAndApplyPatternsOp(
-                    func,
-                    fold_type_extensions_into_contract=True,
-                ).result
-                transform.apply_cse(func)
-                canonicalize(func)
-                payload_mod = apply_registered_pass(payload_mod, "eliminate-empty-tensors")
-                identity_layout = LayoutMapOption.IdentityLayoutMap
-                payload_mod = transform_bufferization.OneShotBufferizeOp(
-                    payload_mod,
-                    allow_return_allocs_from_loops=True,
-                    bufferize_function_boundaries=True,
-                    function_boundary_type_conversion=identity_layout,
-                ).result
-                # fold memref.subviews into vector.transfer_read/write ops
-                payload_mod = apply_registered_pass(payload_mod, "fold-memref-alias-ops")
-                transform.apply_cse(payload_mod)
-                canonicalize(payload_mod)
-                
-                # convert forall to parallel
-                wg_loops = match_and_split(payload_mod, ops={"scf.forall"})
-                for wg_loop in wg_loops:
-                    wg_loop = loop.loop_forall_to_parallel([anytype], wg_loop)
-                func = transform.get_parent_op(anytype, wg_loop)
-                # convert to scf.parallel to gpu.launch
-                func = apply_registered_pass(func, "gpu-map-parallel-loops")
-                func = apply_registered_pass(func, "convert-parallel-loops-to-gpu")
-                func = apply_registered_pass(func, "lower-affine")
-                transform.apply_cse(func)
-                canonicalize(func)
-                
-                # set the number of threads for the gpu.launch operation
-                launch_op = match_and_split(func, ops={"gpu.launch"})
-                num_threads = parameters["sg_rows"] * parameters["subgroup_size"]
-                xegpu.set_gpu_launch_threads(launch_op[0], threads=[num_threads, 1, 1])
-                
-                # outline gpu func
-                func = apply_registered_pass(func, "lower-affine")
-                canonicalize(func)
-                func = apply_registered_pass(func, "gpu-launch-sink-index-computations")
-                payload_mod = apply_registered_pass(payload_mod, "gpu-kernel-outlining")
-                transform.apply_cse(payload_mod)
-                
-                # set xevm target
-                payload_mod = apply_registered_pass(
-                    payload_mod,
-                    "xevm-attach-target",
-                    options={"O": "3", "chip": "bmg"},
-                )
-
-                # convert vector to xegpu
-                gpu_mod_ops = match_and_split(payload_mod, ops={"gpu.module"})
-                for gpu_mod in gpu_mod_ops:
-                    gpu_func = match(gpu_mod, ops={"gpu.func"})
-                    gpu_func = apply_registered_pass(gpu_func, "convert-vector-to-xegpu")
-                    transform.apply_cse(gpu_func)
-                    
-                # Set layout attributes for xegpu.store_nd operations
-                store_ops = match_and_split(gpu_func, ops={"xegpu.store_nd"}, nhandles=1)
-                # for store_op in store_ops:
-                xegpu.set_op_layout_attr(store_ops[0], sg_layout=[8, 1], sg_data=[8, 64])
-                
-                payload_mod = apply_registered_pass(
-                    payload_mod, "gpu-lower-to-xevm-pipeline", options={"xegpu-op-level": "workgroup"}
-                )
-                # Required: yield to end the transform sequence
-                transform.yield_()
-        
-        return [get_bench_wrapper_schedule(self), mod]
+        """Generate transform schedule for softmax."""
+        return [
+            get_bench_wrapper_schedule(self),
+            get_softmax_schedule_module(
+                stop_at_stage=stop_at_stage,
+                parameters=parameters,
+            ),
+        ]
 
     def shared_libs(self) -> list[str]:
         return ["libmlir_levelzero_runtime.so"]
